solar2ev/requirements/solar2ev_req.txt
beautifulsoup4
calplot
#keras==2.10.0
keras_tuner
matplotlib
#numpy==1.22.3
pandas
#rdt==1.3.0
requests
scikit_learn
scipy

#sdmetrics==0.9.1
#tensorflow==2.15.0.post1
#tensorflow_intel==2.11.0

tensorflow==2.10

blynk-library-python

sdv
sdgym
anyio

prettytable
seaborn
secret
tflite_support

pydot
pydotplus
graphviz
openpyxl

kaleido

# AND choco install graphviz from conda env WITH admin capability

